[5779] Amended specs for the pages
